requirements.txt
streamlit
pymupdf
pytesseract
Pillow
langchain
openai
sentence-transformers
faiss-cpu
python-dotenv
